mail.mem.Accounts implementation

+++ src/main/java/org/luwrain/pim/mail/MailAccounts.java
@@ -24,7 +24,7 @@ public interface MailAccounts
     int getId(StoredMailAccount account) throws PimException;
     StoredMailAccount[] load() throws PimException;
     String getUniRef(StoredMailAccount account) throws PimException;
-    StoredMailAccount loadById(long id) throws PimException;
+    StoredMailAccount loadById(int id) throws PimException;
     StoredMailAccount loadByUniRef(String uniRef) throws PimException;
     void save(MailAccount account) throws PimException;
     void delete(StoredMailAccount account) throws PimException;
